Update requirements.txt
I have a suggestion regarding the setup of our virtual environment. Since we are using the transformers and dotenv packages, I propose that we include them in our requirements.txt file to streamline the environment setup process. If you think these additions are unnecessary, please feel free to reject this suggestion(*'v' *)